Fix resource deletion

+++ src/main/kotlin/de/mineking/manager/data/Resource.kt
@@ -33,9 +33,13 @@ interface ResourceTable<T : Resource> : IdentifiableTable<T> {
 	}
 
 	override fun delete(id: String): Boolean {
-		main.participants.delete(Where.equals("parent", id))
-		File("files/$id").deleteRecursively()
+		val result = super.delete(id)
 
-		return super.delete(id)
+		if(result) {
+			main.participants.delete(Where.equals("parent", id))
+			File("files/$id").deleteRecursively()
+		}
+
+		return result
 	}
 }
